Improved decryption.
Working on #2. After some brief testing, it seems the
encryption/decryption works for both txt and jpg files. Currently the
entire process is repeated several times to add layers of encryption,
but these may not actually do anything. Also, there may be a
vulnerability involving being able to almost recreate a file with a key
that's almost the same as the correct one.