Created a GUI for Scrambler.
Working on #3. Moved the code for scrambler to a new file, and modified
scramble and unscramble methods to use file names as parameters. Added
appropriate methods to help with this. Deleted the previous main method
since a GUI will be used from now on. Still need to improve GUI, but
it is functional now.