Made multiple GUI improvments.
Working on #3. Fixed an issue with error messages not having enough
space to display. Also anchored the error message label. Added a button that, when clicked, auto fills the
output file directory based on the input file directory. Created a
confirmation box that pops up when an output file with the given name
already exists.